feat: Limit number of calls for `oracleRoundState`
