+++ core/src/worker/data-feed.utils.ts
@@ -7,32 +7,6 @@ import { CHAIN, DATA_FEED_SERVICE_NAME } from '../settings'
 import { Aggregator__factory } from '@bisonai/orakl-contracts'
 
 /**
- * Get address of node operator given an `oracleAddress`. The data are fetched from the Orakl Network API.
- *
- * @param {string} oracle address
- * @return {string} address of node operator
- * @exception {OraklErrorCode.GetReporterRequestFailed} raises when request failed
- */
-export async function getOperatorAddress({
-  oracleAddress,
-  logger
-}: {
-  oracleAddress: string
-  logger: Logger
-}) {
-  logger.debug('getOperatorAddress')
-
-  return await (
-    await getReporterByOracleAddress({
-      service: DATA_FEED_SERVICE_NAME,
-      chain: CHAIN,
-      oracleAddress,
-      logger
-    })
-  ).address
-}
-
-/**
  * Compute the number of seconds until the next round.
  *
  * FIXME modify aggregator to use single contract call
